Migration

+++ database/migrations/2024_11_03_161959_create_payments_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->timestamps();
 
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('course_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('course_id');
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->foreign('course_id')->references('id')->on('courses')->onDelete('cascade');
