+++ database/migrations/2024_11_09_150514_chapter_course.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('chapter_course', function (Blueprint $table) {
             $table->id();
 
-            $table->unsignedInteger('chapter_id');
-            $table->unsignedInteger('course_id');
+            $table->unsignedBigInteger('chapter_id');
+            $table->unsignedBigInteger('course_id');
 
             $table->foreign('chapter_id')->references('id')->on('chapters')->onDelete('cascade');
             $table->foreign('course_id')->references('id')->on('courses')->onDelete('cascade');
